chat utility on simple chat api
